+++ modules/main/src/main/java/com/nisovin/shopkeepers/util/bukkit/DataUtils.java
@@ -6,6 +6,7 @@ import org.checkerframework.checker.nullness.qual.Nullable;
 
 import com.nisovin.shopkeepers.api.internal.util.Unsafe;
 import com.nisovin.shopkeepers.api.util.UnmodifiableItemStack;
+import com.nisovin.shopkeepers.util.annotations.ReadOnly;
 import com.nisovin.shopkeepers.util.data.container.DataContainer;
 import com.nisovin.shopkeepers.util.inventory.ItemUtils;
 
@@ -34,8 +35,9 @@ public final class DataUtils {
 		return ItemUtils.shallowCopy(itemStack);
 	}
 
-	// Additional processing whenever we load deserialized item stacks.
-	public static @Nullable ItemStack deserializeItemStack(@Nullable ItemStack loadedItemStack) {
+	// Additional processing whenever we load serialized item stacks.
+	// ReadOnly: Returns a new ItemStack instance if modifications are necessary.
+	public static @Nullable ItemStack deserializeItemStack(@ReadOnly @Nullable ItemStack loadedItemStack) {
 		if (loadedItemStack == null) return null;
 		// Note: Spigot creates Bukkit ItemStacks, whereas Paper automatically replaces the
 		// deserialized Bukkit ItemStacks with CraftItemStacks. However, as long as the deserialized
@@ -57,7 +59,7 @@ public final class DataUtils {
 		return processed;
 	}
 
-	public static ItemStack deserializeNonNullItemStack(ItemStack loadedItemStack) {
+	public static ItemStack deserializeNonNullItemStack(@ReadOnly ItemStack loadedItemStack) {
 		return Unsafe.assertNonNull(deserializeItemStack(loadedItemStack));
 	}